Slylandro Probes - Spoken coordinates, 1st pass.
Moves the generation of coordinates into a callback func to make
things easier to understand/work with (for me, at least) and open
up different uses for the GLOBAL_PLAYER_LOCATION phrase hook

Uses content tricks to get page breaks to happen.  Ugly, and not
perfect, as the coordinates themselves are broken up.  Need to fix.